Feature updates for 29112020 - deleted unwanted files for controllers - deleted unwanted files for services - created UserController.java and UserService.java for single point entry - re arranged file imports and dependencies
