+++ Trello_quora/quora-service/src/main/java/com/upgrad/quora/service/entity/UserAuthEntity.java
@@ -6,6 +6,11 @@ import java.time.ZonedDateTime;
 
 @Entity
 @Table(name = "user_auth")
+@NamedQueries({
+  @NamedQuery(
+      name = "fetchThisUserByJWTAuthToken",
+      query = "select u from UserAuthEntity u where u.accessToken=:accessToken")
+})
 public class UserAuthEntity {
 
     @Id
